Bug:21938
Now the default of option fastload is changed to not set.
if option fastload is set to either fast or false, it will be honored.
if output_directory is set, it will be honored. Assuming output_directory is set, if option fastload is not set, fastload mode is on.

For this bug, as -output_partition is not set while server uses range policy, mlcp can't run in fastload mode. Therefore, the correct command is to specify -fastload false -output_directory /medline. In this way mlcp will honor the -output_directory and run in non-fast load mode.