Add command-line parameters to support passing a KeyStore
MLCP 9.0-1 added support for TLS connections by passing the '-ssl'
parameter on the command line and later controlling the TLS protocol
using the '-ssl_protocol' parameter. However both of these only allow
a TLS Server Authentication connection, in order to make a connection
to a MarkLogic server that requires TLS Client Authentication as well
as Server Authentication a SSLConfigOptions class needs to be written
and the requisite SSL Configuration passed to MLCP using the '-conf'
parameter.

https://docs.marklogic.com/guide/mlcp/install#id_62768

TLS Client Authentication is now quite a common requirement rather
than an 'Advanced' feature and requiring a user to have the necessary
skills to code an SSLConfigOptions class to handle a KeyManager makes
it unnecessarily difficult to use for a lot of users.

This enhancement allows the user to pass an existing Java KeyStore
and/or TrustStore via the MLCP command-line to make a TLS Client
Authentication connection without the need to code a SSLConfigOptions
class.